Changed mechanism to skip deployment of SWT modules during release.
Previously we attempted to skip deploy using a release profile within
the module itself, although this appeared not to be activated when the
release was performed at the top-level.

Instead gone for the simpler approach of skipping all deployment for SWT
modules.  This does have the drawback of not being able to deploy SWT
modules outside of a release.